refactor: move read_and_validate methods to validation
