+++ src/core/structure/modules/imports.rs
@@ -1,9 +1,5 @@
 use crate::{
-    core::{
-        decoding::reader::WasmReader,
-        structure::modules::indices::{IdxVec, TypeIdx},
-    },
-    DecodingError, ExternType, FuncType, GlobalType, MemType, TableType, ValidationError,
+    core::structure::modules::indices::TypeIdx, ExternType, GlobalType, MemType, TableType,
     ValidationInfo,
 };
 
@@ -14,47 +10,12 @@ pub struct Import<'wasm> {
     pub desc: ImportDesc,
 }
 
-impl<'wasm> Import<'wasm> {
-    pub fn read_and_validate(
-        wasm: &mut WasmReader<'wasm>,
-        c_types: &IdxVec<TypeIdx, FuncType>,
-    ) -> Result<Self, ValidationError> {
-        let module_name = wasm.read_name()?;
-        let name = wasm.read_name()?;
-        let desc = ImportDesc::read_and_validate(wasm, c_types)?;
-
-        Ok(Self {
-            module_name,
-            name,
-            desc,
-        })
-    }
-}
-
 #[derive(Debug, Clone)]
 pub enum ImportDesc {
     Func(TypeIdx),
     Table(TableType),
     Mem(MemType),
     Global(GlobalType),
-}
-
-impl ImportDesc {
-    pub fn read_and_validate(
-        wasm: &mut WasmReader,
-        c_types: &IdxVec<TypeIdx, FuncType>,
-    ) -> Result<Self, ValidationError> {
-        let desc = match wasm.read_u8()? {
-            0x00 => Self::Func(TypeIdx::read_and_validate(wasm, c_types)?),
-            // https://webassembly.github.io/spec/core/binary/types.html#table-types
-            0x01 => Self::Table(TableType::read(wasm)?),
-            0x02 => Self::Mem(MemType::read(wasm)?),
-            0x03 => Self::Global(GlobalType::read(wasm)?),
-            other => return Err(DecodingError::MalformedImportDescDiscriminator(other).into()),
-        };
-
-        Ok(desc)
-    }
 }
 
 impl ImportDesc {
