+++ src/validation/modules/element_segments.rs
@@ -1,0 +1,291 @@
+use alloc::{collections::btree_set::BTreeSet, vec::Vec};
+
+use crate::{
+    core::{
+        decoding::{
+            modules::parse_elemkind,
+            reader::{span::Span, WasmReader},
+        },
+        structure::modules::{
+            element_segments::{ActiveElem, ElemItems, ElemMode, ElemType},
+            indices::{FuncIdx, IdxVec, TableIdx, TypeIdx},
+        },
+    },
+    read_constant_expression::read_constant_expression,
+    validation_stack::ValidationStack,
+    GlobalType, NumType, RefType, TableType, ValType, ValidationError,
+};
+
+impl ElemType {
+    // TODO: @nerodesu017 maybe split up the validation from the parsing?
+    /// Here we can't implement WasmReadable because we also want a mutable
+    /// reference to a [`BTreeSet<u32>`] (`referenced_functions`)
+    ///
+    /// This comes in handy later on when we are validating the actual code of
+    /// the functions so that we can make sure we are not referencing invalid
+    /// functions
+    pub fn read_and_validate(
+        wasm: &mut WasmReader,
+        c_funcs: &IdxVec<FuncIdx, TypeIdx>,
+        validation_context_refs: &mut BTreeSet<FuncIdx>,
+        c_tables: &IdxVec<TableIdx, TableType>,
+        imported_global_types: &[GlobalType],
+    ) -> Result<Vec<Self>, ValidationError> {
+        wasm.read_vec(|wasm| {
+            let prop = wasm.read_var_u32()?;
+
+            let elem = match prop {
+                0 => {
+                    // binary format is: 0:u32 e:expr y*:vec(funcidx)
+                    // should parse to spec struct {type funcref, init ((ref.func y) end)*, mode active {table 0, offset e}}
+                    // which is equivalent to ElemType{init: ElemItems::RefFuncs(y*), mode: ElemMode::Active{0, e}} here
+                    let e = parse_validate_active_segment_offset_expr(
+                        wasm,
+                        imported_global_types,
+                        c_funcs,
+                        validation_context_refs,
+                    )?;
+                    let init = parse_validate_shortened_initializer_list(
+                        wasm,
+                        c_funcs,
+                        validation_context_refs,
+                    )?;
+                    let mode = ElemMode::Active(ActiveElem {
+                        table_idx: TableIdx::validate(0, c_tables)?,
+                        init_expr: e,
+                    });
+                    ElemType { init, mode }
+                }
+                1 => {
+                    // binary format is: 1:u32 et:elemkind y*:vec(funcidx)
+                    // should parse to spec struct {type et, init ((ref.func y) end)*, mode passive}
+                    // which is equivalent to ElemType{init: ElemItems::RefFuncs(y*), mode: ElemMode::Passive} here
+                    let _et = parse_elemkind(wasm)?;
+                    let init = parse_validate_shortened_initializer_list(
+                        wasm,
+                        c_funcs,
+                        validation_context_refs,
+                    )?;
+                    let mode = ElemMode::Passive;
+                    ElemType { init, mode }
+                }
+                2 => {
+                    // binary format is: 2:u32 x:tableidx e:expr et:elemkind y*:vec(funcidx)
+                    // should parse to spec struct {type et, init ((ref.func y) end)*, mode active {table x, offset e}}
+                    // which reflects to ElemType{init: ElemItems::RefFuncs(y*), mode: ElemMode::Active{x, e}} here
+                    let x = TableIdx::read_and_validate(wasm, c_tables)?;
+                    let e = parse_validate_active_segment_offset_expr(
+                        wasm,
+                        imported_global_types,
+                        c_funcs,
+                        validation_context_refs,
+                    )?;
+                    let _et = parse_elemkind(wasm)?;
+                    let init = parse_validate_shortened_initializer_list(
+                        wasm,
+                        c_funcs,
+                        validation_context_refs,
+                    )?;
+                    let mode = ElemMode::Active(ActiveElem {
+                        table_idx: x,
+                        init_expr: e,
+                    });
+                    ElemType { init, mode }
+                }
+                3 => {
+                    // binary format is: 3:u32 et:elemkind y*:vec(funcidx)
+                    // should parse to spec struct {type et, init ((ref.func y) end)*, mode declarative}
+                    // which is equivalent to ElemType{init: ElemItems::RefFuncs(y*), mode: ElemMode::Declarative} here
+                    let _et = parse_elemkind(wasm)?;
+                    let init = parse_validate_shortened_initializer_list(
+                        wasm,
+                        c_funcs,
+                        validation_context_refs,
+                    )?;
+                    let mode = ElemMode::Declarative;
+                    ElemType { init, mode }
+                }
+                4 => {
+                    // binary format is: 4:u32 e:expr el*:vec(expr)
+                    // should parse to spec struct {type funcref, init el*, mode active { table 0, offset e}}
+                    // which is equivalent to ElemType{init: ElemItems::Exprs(funcref, el*), mode: ElemMode::Active{0, e}}
+                    let e = parse_validate_active_segment_offset_expr(
+                        wasm,
+                        imported_global_types,
+                        c_funcs,
+                        validation_context_refs,
+                    )?;
+                    let init = parse_validate_generic_initializer_list(
+                        wasm,
+                        RefType::FuncRef,
+                        imported_global_types,
+                        c_funcs,
+                        validation_context_refs,
+                    )?;
+                    let mode = ElemMode::Active(ActiveElem {
+                        table_idx: TableIdx::validate(0, c_tables)?,
+                        init_expr: e,
+                    });
+                    ElemType { init, mode }
+                }
+                5 => {
+                    // binary format is 5:u32 et: reftype el*:vec(expr)
+                    // should parse to spec struct {type et, init el*, mode passive}
+                    // which is equivalent to ElemType{init: ElemItems::Exprs(et, el*), mode: ElemMode::Passive} here
+                    let et = RefType::read(wasm)?;
+                    let init = parse_validate_generic_initializer_list(
+                        wasm,
+                        et,
+                        imported_global_types,
+                        c_funcs,
+                        validation_context_refs,
+                    )?;
+                    let mode = ElemMode::Passive;
+                    ElemType { init, mode }
+                }
+                6 => {
+                    // binary format is 6:u32 x:table_idx e:expr et:reftype el*:vec(expr)
+                    // should parse to spec struct {type et, init el*, mode passive}
+                    // which is equivalent to ElemType{init: Exprs(et, el*), mode: ElemMode::Active{table x, offset e}} here
+                    let x = TableIdx::read_and_validate(wasm, c_tables)?;
+                    let e = parse_validate_active_segment_offset_expr(
+                        wasm,
+                        imported_global_types,
+                        c_funcs,
+                        validation_context_refs,
+                    )?;
+                    let et = RefType::read(wasm)?;
+                    let init = parse_validate_generic_initializer_list(
+                        wasm,
+                        et,
+                        imported_global_types,
+                        c_funcs,
+                        validation_context_refs,
+                    )?;
+                    let mode = ElemMode::Active(ActiveElem {
+                        table_idx: x,
+                        init_expr: e,
+                    });
+                    ElemType { init, mode }
+                }
+                7 => {
+                    // binary format is 7:u32 et:reftype el*:vec(expr)
+                    // should parse to spec struct {type et, init el*, mode declarative}
+                    // which is equivalent to ElemType{init: Exprs(et, el*), mode: ElemMode::Declarative} here
+                    let et = RefType::read(wasm)?;
+                    let init = parse_validate_generic_initializer_list(
+                        wasm,
+                        et,
+                        imported_global_types,
+                        c_funcs,
+                        validation_context_refs,
+                    )?;
+                    let mode = ElemMode::Declarative;
+                    ElemType { init, mode }
+                }
+                invalid_mode @ 8.. => {
+                    return Err(ValidationError::InvalidElementMode(invalid_mode));
+                }
+            };
+
+            // assume the element segment is well formed in terms of abstract syntax from now on.
+            // start validating element segment of form {type t, init e*, mode elemmode}: https://webassembly.github.io/spec/core/valid/modules.html#element-segments
+            let t = elem.ty();
+            // 1. Each e_i must be valid with type t and be const: this is already checked during the parse of initializer expressions above.
+            // 2. elemmode must be valid with type t
+            // -- start validating elemmode for type t:
+            match elem.mode {
+                ElemMode::Active(ActiveElem {
+                    table_idx: x,
+                    init_expr: _expr,
+                }) => {
+                    // start validating elemmode of form active {table x, offset expr}
+                    // 1-2. C.tables[x] must be defined with type: limits t
+                    // SAFETY: The `ActiveElem` that is being deconstructed was
+                    // created and also validated in this function.
+                    let table_type = unsafe { c_tables.get(x) };
+                    if table_type.et != t {
+                        return Err(ValidationError::ActiveElementSegmentTypeMismatch);
+                    }
+                    // 3-4. _expr must be valid with type I32 and be const: already checked during the parse of initializer expressions above.
+                    // Then elemmode is valid with type t.
+                }
+                ElemMode::Declarative | ElemMode::Passive => (), // these are valid for any type t.
+            }
+            // -- Then elemmmode is valid with type t.
+            // Then the element segment is valid with type t.
+            Ok(elem)
+        })
+    }
+}
+
+// helpers for avoiding code duplication during parsing/validation of element segment
+// https://webassembly.github.io/spec/core/binary/modules.html#element-section
+
+/// Parse and validate an active segment offset.
+/// An active segment offset is valid if its offset expr is of type I32 and is a const expr.
+/// Additionally inserts new items to the set validation_context_refs.
+/// Validation_context_refs corresponds to C.refs in <https://webassembly.github.io/spec/core/valid/conventions.html#context>
+///
+/// # Returns
+/// - `Ok(Span)` of the expr validated to be of type I32 if parsing & validating succeeds, `Err(_)` otherwise.
+fn parse_validate_active_segment_offset_expr(
+    wasm: &mut WasmReader,
+    imported_global_types: &[GlobalType],
+    c_funcs: &IdxVec<FuncIdx, TypeIdx>,
+    validation_context_refs: &mut BTreeSet<FuncIdx>,
+) -> Result<Span, ValidationError> {
+    let mut valid_stack = ValidationStack::new();
+    let (span, seen_func_refs) =
+        read_constant_expression(wasm, &mut valid_stack, imported_global_types, c_funcs)?;
+    validation_context_refs.extend(seen_func_refs);
+    valid_stack.assert_val_types(&[ValType::NumType(NumType::I32)], true)?;
+    Ok(span)
+}
+
+/// Parse and validate a vector of func_idx's that reflect as the initializer list of an element segment in the form of ((ref.func func_idx) end) of in the abstract syntax.
+/// An expression of such form is valid and const with type funcref if `C.funcs[func_idx]` exists (see link for the definition of validation context)
+/// This codebase holds these as `ElemItems::RefFuncs(Vec<u32>)` to sidestep the need to show `Span` for each expression.
+/// Additionally inserts new items to the set validation_context_refs.
+/// validation_context_refs corresponds to C.refs in <https://webassembly.github.io/spec/core/valid/conventions.html#context>
+///
+/// # Returns
+/// - `Ok(ElemItems::RefFuncs(_))` corresponding to the parsed list if parsing & validating succeeds, `Err(_)` otherwise.
+fn parse_validate_shortened_initializer_list(
+    wasm: &mut WasmReader,
+    c_funcs: &IdxVec<FuncIdx, TypeIdx>,
+    validation_context_refs: &mut BTreeSet<FuncIdx>,
+) -> Result<ElemItems, ValidationError> {
+    wasm.read_vec(|w| {
+        let func_idx = FuncIdx::read_and_validate(w, c_funcs)?;
+        validation_context_refs.insert(func_idx);
+        Ok(func_idx)
+    })
+    .map(ElemItems::RefFuncs)
+}
+
+/// Parse and validate the initializer list of an element segment for the supplied type `expected_type`.
+/// An initializer list is valid with type `expected_type` if all of the expressions within is const and is of type `expected_type`.
+/// This codebase holds these as `ElemItems::Exprs(RefType, Vec<Span>)`.
+/// Additionally inserts new items to the set validation_context_refs.
+/// validation_context_refs corresponds to C.refs in <https://webassembly.github.io/spec/core/valid/conventions.html#context>
+///
+/// # Returns
+/// - `Ok(ElemItems::Exprs(expected_type, _))` corresponding to the parsed list if parsing & validating succeeds, `Err(_)` otherwise.
+fn parse_validate_generic_initializer_list(
+    wasm: &mut WasmReader,
+    expected_type: RefType,
+    imported_global_types: &[GlobalType],
+    c_funcs: &IdxVec<FuncIdx, TypeIdx>,
+    validation_context_refs: &mut BTreeSet<FuncIdx>,
+) -> Result<ElemItems, ValidationError> {
+    wasm.read_vec(|w| {
+        let mut valid_stack = ValidationStack::new();
+        let (span, seen_func_refs) =
+            read_constant_expression(w, &mut valid_stack, imported_global_types, c_funcs)?;
+        validation_context_refs.extend(seen_func_refs);
+        valid_stack.assert_val_types(&[ValType::RefType(expected_type)], true)?;
+        Ok(span)
+    })
+    .map(|v| ElemItems::Exprs(expected_type, v))
+}
