+++ src/validation/modules/mod.rs
@@ -1,0 +1,4 @@
+pub mod element_segments;
+pub mod exports;
+pub mod imports;
+pub mod indices;
